chore: components 디렉토리의 파일 views로 이동(FE)
