+++ .gitignore
@@ -39,6 +39,7 @@ out/
 
 ### VS Code ###
 .vscode/
+/node_modules/
 
 
 ## DB 정보##
