+++ frontend/.gitignore.txt
@@ -1,5 +1,5 @@
 .DS_Store
-node_modules
+/node_modules/
 /dist
 
 # local env files
