--- frontend/src/components/Signuppage.vue
@@ -1,172 +1,0 @@
-<template>
-  <v-card class="mx-auto" style="max-width: 500px">
-    <v-toolbar color="deep-purple-accent-4" cards dark flat>
-      <v-btn icon>
-        <v-icon>mdi-arrow-left</v-icon>
-      </v-btn>
-      <v-card-title class="text-h6 font-weight-regular"> Sign up </v-card-title>
-      <v-spacer></v-spacer>
-      <v-btn icon>
-        <v-icon>mdi-magnify</v-icon>
-      </v-btn>
-      <v-btn icon>
-        <v-icon>mdi-dots-vertical</v-icon>
-      </v-btn>
-    </v-toolbar>
-    <v-form ref="form" v-model="isValid" class="pa-4 pt-6">
-      <v-text-field
-        v-model="password"
-        :rules="[rules.password, rules.length(6)]"
-        variant="filled"
-        color="deep-purple"
-        counter="6"
-        label="Password"
-        style="min-height: 96px"
-        type="password"
-        id="password" 
-        name="password"
-      ></v-text-field>
-      
-
-      <v-text-field
-        v-model="name"
-        variant="filled"
-        color="deep-purple"
-        label="Name"
-        id="name"
-      ></v-text-field>
-
-      <v-text-field
-        v-model="email"
-        :rules="[rules.email]"
-        variant="filled"
-        color="deep-purple"
-        label="Email address"
-        type="email"
-        id="email"
-      ></v-text-field>
-
-      <v-textarea
-        v-model="bio"
-        auto-grow
-        variant="filled"
-        color="deep-purple"
-        label="Bio"
-        rows="1"
-      ></v-textarea>
-      <v-checkbox
-        v-model="agreement"
-        :rules="[rules.required]"
-        color="deep-purple"
-      >
-        <template v-slot:label>
-          I agree to the&nbsp;
-          <a href="#" @click.stop.prevent="dialog = true">Terms of Service</a>
-          &nbsp;and&nbsp;
-          <a href="#" @click.stop.prevent="dialog = true">Privacy Policy</a>*
-        </template>
-      </v-checkbox>
-    </v-form>
-    <v-divider></v-divider>
-    <v-card-actions>
-      <v-btn variant="text" @click="form.reset()"> Clear </v-btn>
-      <v-spacer></v-spacer>
-      <v-btn @click="submitForm" :disabled="!isValid" :loading="isLoading" color="deep-purple-accent-4">
-  제출
-</v-btn>
-    </v-card-actions>
-    <v-dialog v-model="dialog" max-width="400" persistent>
-      <v-card>
-        <v-card-title class="text-h5 bg-grey-lighten-3"> Legal </v-card-title>
-        <v-card-text>
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
-          eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
-          minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-          aliquip ex ea commodo consequat. Duis aute irure dolor in
-          reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-          pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-          culpa qui officia deserunt mollit anim id est laborum.
-        </v-card-text>
-        <v-divider></v-divider>
-        <v-card-actions>
-          <v-btn variant="text" @click="(agreement = false), (dialog = false)">
-            No
-          </v-btn>
-          <v-spacer></v-spacer>
-          <v-btn
-            color="deep-purple"
-            variant="tonal"
-            @click="(agreement = true), (dialog = false)"
-          >
-            Yes
-          </v-btn>
-        </v-card-actions>
-      </v-card>
-    </v-dialog>
-  </v-card>
-</template>
-<script>
-import axios from 'axios';
-
-
-axios.defaults.withCredentials = true;
-
-
-
-
-export default {
-  data: () => ({
-    name: undefined,
-
-    agreement: false,
-    bio: "Far far away, behind the word mountains, far from the countries Vokalia and Consonantia, there live the blind texts",
-    dialog: false,
-    email: undefined,
-    isValid: false,
-    isLoading: false,
-    password: undefined,
-    phone: undefined,
-    rules: {
-      email: (v) => !!(v || "").match(/@/) || "Please enter a valid email",
-      length: (len) => (v) =>
-        (v || "").length >= len || `Invalid character length, required ${len}`,
-      password: (v) =>
-        !!(v || "").match(
-          /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*(_|[^\w])).+$/
-        ) ||
-        "Password must contain an upper case letter, a numeric character, and a special character",
-      required: (v) => !!v || "This field is required",
-
- },
-  }),
-
-methods: {
-    submitForm() {
-      const request_data = {
-        memberEmail: this.email,
-        memberPw: this.password,
-        memberName: this.name,
-        memberId: "1111",
-        Status : true
-
-      };
-
-      console.log(request_data);
-
-      axios.post("http://localhost:8080/member/register", request_data)
-        .then(() => {
-          console.log("######### success");
-          location.href = "./mainPage.html"; //나중에 로그인 페이지로 넘겨라
-        })
-        .catch(error => {
-          console.error("Error submitting form:", error);
-        });
-    },
-  },
-};
-</script>
-
-
-
-
-
